+++ src/Screens/RegistrationScreen.js
@@ -12,7 +12,7 @@ import {
   TouchableWithoutFeedback,
   View,
 } from "react-native";
-import backgroundImg from "..//images/bg-pic.jpg";
+import backgroundImg from "../../assets/images/bg-pic.jpg";
 import { useState } from "react";
 import Icon from "react-native-vector-icons/Ionicons";
 import { useNavigation } from "@react-navigation/native";
