+++ src/Screens/index.js
@@ -1,0 +1,7 @@
+export * from "./RegistrationScreen";
+export * from "./LoginScreen";
+export * from "./PostsScreen";
+export * from "./CreatePostScreen";
+export * from "./ProfileScreen";
+export * from ".//CommentsScreen";
+export * from "./MapScreen";
