+++ src/components/CameraView.js
