+++ src/components/index.js
@@ -1,0 +1,5 @@
+export * from "./BackBtn";
+export * from "./LogOutBtn";
+export * from "./CameraView";
+export * from "./Post";
+export * from "./PostProfile";
